crew/requirements.txt
crewai[tools]
typing
fastapi
uvicorn
python-dotenv
pydantic